test(coverage): Find likely important py3 holes
In order to focus the porting work it's helpful to have a visual representation of
coverage holes.  For example, so I can see what uncovered parts of
`./src/allmydata/node.py` are most likley to be both important and need porting work.

Since CI isn't using `./Makefile` for Windows compatibility and the XML report tends not
to be useful for local development (AFAIK), might as well just replace it with the HTML
report locally.