feat(py3): Fix test_introducer bytes literals
This is really a class of errors throughout the tests where a test is passing in bytes
under py2 via bare string literals that are now unicode under py3.  IMO, it would be
fruitful to do all the clear and easy cases of this in one PR.

I surveyed the use of `remove_prefix()` throughout the code base and most of the uses
seem to explicitly use byte strings so I think switching these instances to do the same
makes the most sense.