feat(refs #3455): Fix PY3 UTF-8 BOM node config
Update the handling of UTF-8 BOM magic number characters under Python 3 in both the real
function and the test.

To ensure that the test covers the case of a file with the right character, I stopped
the test before these changes in the debugger, copied the node configuration file the
test generates with the BOM magic number to the test data and then switched the test to
use that file rather than re-creating it.  IMO this is more reliable coverage of the
condition being tested as it prevents something in the generation of the
file (e.g. running under a newer Python version) resulting in a false passing result.